bugFix:
BarButtons on bottom Toolbar now fire.  Had to extend backgroundImageView to the bottom of view.safeareaLayout.Bottom.  Before it has a constraint of -45

Looks like NSLayoutConstraint pushed the toolbar out of the area that backgroundImageView occupies.  So even though the imageView has ".isUserInteraction = true".  The toolbar was no longer capable of interacting with user.